feat: implemented multiple new capabilities
* periodicity
realtime days are now broken into 1.) multiple in game trading days
and 2.) each in-game trading day is broken up into a specified number
of periods.

* search functionality
rather than creating a single websocket request for each ticker,
clients can pass in a vector of tickers, and they will be filtered
from the companies data and emitted as a vector.